Create a new file for lobby commands
Grab the one existing command and put it in there as well.
