+++ src/commands/all.js
@@ -2,6 +2,7 @@ import './edge.js'
 import './exchange.js'
 import './help.js'
 import './keys.js'
+import './lobby.js'
 import './login.js'
 import './password.js'
 import './pin.js'
